more batch support
