+++ drive/src/drive/identity.rs
@@ -24,6 +24,9 @@ impl Drive {
             )),
             &mut insert_operations,
         )?;
+        if apply {
+
+        }
         calculate_fee(None, None, Some(insert_operations), None)
     }
